backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
gunicorn==21.2.0
python-dotenv==1.0.0
openpyxl==3.1.5
azure-storage-blob==12.19.0
pandas==2.1.4
numpy==1.26.3
